[UBA] * Refactored out uba horde integration to its own module called UbaCoordinatorHorde. This required some changes like callback registration to register helpers to uba host * Added UbaCoordinatorHorde target in uba to be able to compile a coordinator for horde that can be used in UbaCli or other custom written uba tools
#rb laura.hermanns

[CL 32347975 by henrik karlsson in ue5-main branch]

+++ Engine/Source/Developer/UbaCoordinatorHorde/Private/UbaHordeMetaClient.cpp
@@ -134,7 +134,7 @@ TSharedPtr<FUbaHordeMetaClient::HordeMachinePromise, ESPMode::ThreadSafe> FUbaHo
 					}
 					if (PropertyElementString.StartsWith(TEXT("LogicalCores=")))
 					{
-						Info.LogicalCores = FCString::Atoi(*PropertyElementString + 13);
+						Info.LogicalCores = (uint16)FCString::Atoi(*PropertyElementString + 13);
 					}
 				}
 			}
